# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Documentación de la API de Desastres",
        "description": "Esta es la documentación de la API del proyecto Desastres.",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/images": {
            "get": {
                "tags": [
                    "Imágenes"
                ],
                "summary": "Listar todas las imágenes",
                "operationId": "2352228068ea2ab96f614dd425ae11af",
                "responses": {
                    "200": {
                        "description": "Lista de imágenes",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Image"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Imágenes"
                ],
                "summary": "Subir una nueva imagen",
                "operationId": "6b7ebc376b155537f1858fc7f3e6038c",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "url",
                                    "note_id"
                                ],
                                "properties": {
                                    "url": {
                                        "type": "string",
                                        "example": "images/image.jpg"
                                    },
                                    "note_id": {
                                        "type": "integer",
                                        "example": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Imagen creada",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Image"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/api/notes": {
            "get": {
                "tags": [
                    "Notas"
                ],
                "summary": "Listar todas las notas",
                "operationId": "9a95bde5931986ad60b9f7f414ec425e",
                "responses": {
                    "200": {
                        "description": "Lista de notas",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Note"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Notas"
                ],
                "summary": "Crear una nueva nota",
                "operationId": "d133c170557b4935898f2e444e0fd418",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "title",
                                    "description",
                                    "ubication",
                                    "status",
                                    "category_id"
                                ],
                                "properties": {
                                    "title": {
                                        "type": "string",
                                        "example": "Incendio en el bosque"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Un gran incendio en el bosque ha causado daños significativos."
                                    },
                                    "ubication": {
                                        "type": "string",
                                        "example": "Bosque Nacional"
                                    },
                                    "status": {
                                        "type": "boolean",
                                        "example": true
                                    },
                                    "category_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "images": {
                                        "type": "array",
                                        "items": {
                                            "type": "string",
                                            "format": "binary"
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Nota creada",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Note"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/notes/{id}": {
            "get": {
                "tags": [
                    "Notas"
                ],
                "summary": "Mostrar una nota específica",
                "operationId": "4f26c25d4040d0ef0fa8d8a5ea9fa4da",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Detalles de la nota",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Note"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Nota no encontrada"
                    }
                }
            },
            "put": {
                "tags": [
                    "Notas"
                ],
                "summary": "Actualizar una nota",
                "operationId": "c7e6224fadd6d82b81ded7615279a15c",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "title",
                                    "description",
                                    "ubication",
                                    "status",
                                    "category_id"
                                ],
                                "properties": {
                                    "title": {
                                        "type": "string",
                                        "example": "Incendio en el bosque actualizado"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Actualización: el incendio ha sido controlado."
                                    },
                                    "ubication": {
                                        "type": "string",
                                        "example": "Bosque Nacional"
                                    },
                                    "status": {
                                        "type": "boolean",
                                        "example": true
                                    },
                                    "category_id": {
                                        "type": "integer",
                                        "example": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Nota actualizada",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Note"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "Notas"
                ],
                "summary": "Eliminar una nota",
                "operationId": "99b63679b2822651f8f0ae585f927f0b",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Nota eliminada",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Se ha eliminado la nota correctamente"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Nota no encontrada"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Image": {
                "required": [
                    "url",
                    "note_id"
                ],
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "url": {
                        "type": "string",
                        "example": "images/image.jpg"
                    },
                    "note_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
            "Note": {
                "required": [
                    "title",
                    "description",
                    "ubication",
                    "status",
                    "category_id"
                ],
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "title": {
                        "type": "string",
                        "example": "Incendio en el bosque"
                    },
                    "description": {
                        "type": "string",
                        "example": "Un gran incendio en el bosque ha causado daños significativos."
                    },
                    "ubication": {
                        "type": "string",
                        "example": "Bosque Nacional"
                    },
                    "status": {
                        "type": "boolean",
                        "example": true
                    },
                    "category_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "user_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "images": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/Image"
                        }
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "sanctum": {
                "type": "apiKey",
                "description": "Ingresa el token en el formato (Bearer <token>)",
                "name": "Authorization",
                "in": "header"
            }
        }
    },
    "security": [
        {
            "sanctum": []
        }
    ]
}